# webserver/app/static/openapi.json
{
  "openapi": "3.0.2",
  "info": {
    "description": "Federated Node",
    "title": "Federated Node",
    "version": "0.0.1"
  },
  "servers":[
    {
      "url": "/",
      "description": "Live service"
    },
    {
      "url": "http://127.0.0.1:5000",
      "description": "Local server"
    }
  ],
...
  "paths":{
...
    "/tasks/{id}": {
      "get": {
        "operationId": "get_task_by_id",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Unique task identifier",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "tags": ["Tasks"],
        "summary": "Get a single task by its id",
        "responses": {
          "200": {
            "$ref": "#/components/responses/Task"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/tasks": {
      "get": {
        "operationId": "get_tasks",
        "tags": ["Tasks"],
        "summary": "Get a full list of tasks",
        "responses": {
          "200": {
            "$ref": "#/components/responses/TaskList"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      },
      "post": {
        "operationId": "create_task",
        "tags": ["Tasks"],
        "summary": "Queue a new analytics task on a given dataset",
        "requestBody": {
          "content": {
            "application/json":{
              "schema":{
                "$ref": "#/components/schemas/TaskPostBody"
              }
            }
          }
        },
        "responses":{
          "201":{
            "$ref": "#/components/responses/TaskPost"
          },
          "400":{
            "$ref": "#/components/responses/InvalidBody"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "409":{
            "$ref": "#/components/responses/Conflict"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/tasks/validate": {
      "post": {
        "operationId": "validate_task",
        "tags": ["Tasks"],
        "summary": "Validate a task request. Checks if the docker image can be found, and the overall required fields to create a task",
        "requestBody": {
          "content": {
            "application/json":{
              "schema":{
                "$ref": "#/components/schemas/TaskPostBody"
              }
            }
          }
        },
        "responses":{
          "201":{
            "$ref": "#/components/responses/SimpleOk"
          },
          "400":{
            "$ref": "#/components/responses/InvalidBody"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/tasks/{id}/cancel": {
      "post": {
        "operationId": "cancel_task",
        "tags": ["Tasks"],
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Unique task identifier",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "summary": "Cancel a running task",
        "responses":{
          "201":{
            "$ref": "#/components/responses/SimpleOk"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/tasks/{id}/results": {
      "get": {
        "operationId": "get_task_results",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Unique task identifier",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "tags": ["Tasks"],
        "summary": "Get task results if available",
        "responses":{
          "200":{
            "$ref": "#/components/responses/TaskResults"
          },
          "400":{
            "$ref": "#/components/responses/ReviewNeededOrBlocked"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/tasks/{id}/results/approve": {
      "post": {
        "operationId": "approve_results",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Unique task identifier",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "tags": ["Tasks"],
        "summary": "Approve the results release",
        "responses":{
          "201":{
            "description": "Review successful"
          },
          "400":{
            "$ref": "#/components/responses/MultipleReviewError"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/tasks/{id}/results/block": {
      "post": {
        "operationId": "block_results",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Unique task identifier",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "tags": ["Tasks"],
        "summary": "Block the results release",
        "responses":{
          "201":{
            "description": "Review successful"
          },
          "400":{
            "$ref": "#/components/responses/MultipleReviewError"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
    "/tasks/{id}/logs": {
      "get": {
        "operationId": "get_task_logs",
        "parameters": [
          {
            "in": "path",
            "name": "id",
            "description": "Unique task identifier",
            "required": true,
            "schema": {"type": "integer"}
          }
        ],
        "tags": ["Tasks"],
        "summary": "Get task logs if available",
        "responses":{
          "200":{
            "$ref": "#/components/responses/TaskLogs"
          },
          "401":{
            "$ref": "#/components/responses/Unauthenticated"
          },
          "403":{
            "$ref": "#/components/responses/Unauthorized"
          },
          "404":{
            "$ref": "#/components/responses/NotFound"
          },
          "500":{
            "$ref": "#/components/responses/InternalError"
          }
        }
      }
    },
...
  },
  "security": [
    {
      "bearerAuth": []
    }
  ],
  "components":{
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
...
    "responses": {
...
      "Task": {
        "description": "Task description",
        "content": {
          "application/json":{
            "schema":{
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TaskById"
              }
            }
          }
        }
      },
      "TaskList": {
        "description": "List of tasks",
        "content": {
          "application/json":{
            "schema":{
              "type": "array",
              "items": {
                "$ref": "#/components/schemas/TaskById"
              }
            }
          }
        }
      },
      "TaskPost": {
        "description": "Successfully added",
        "content": {
          "application/json":{
            "schema":{
              "type": "object",
              "properties": {
                "task_id": {
                  "type": "string",
                  "example": "1"
                }
              }
            }
          }
        }
      },
      "TaskResults": {
        "description": "Successfully added",
        "content": {
          "application/zip":{
            "schema":{
              "type": "string",
              "format": "binary"
            }
          }
        }
      },
      "TaskLogs": {
        "description": "Pod's logs",
        "content": {
          "application/json":{
            "schema":{
              "type":"object",
              "properties": {
                "logs":{
                  "oneOf": [
                    {
                      "type": "array",
                      "items": {
                        "type": "string",
                        "example": "2025-03-11T14:14:43.770925327Z R version 4.3.3 (2024-02-29) -- \"Angel Food Cake\""
                      }
                    },{
                      "type": "string",
                      "description": "Response when a task is still initializing",
                      "example": "Task queued"
                    }
                  ]
                }
              }
            }
          }
        }
      },
...
      "SimpleOk": {
        "description": "Simple text ok response",
        "content": {
          "plain-text":{
            "example": "Ok"
          }
        }
      },
...
      "ReviewNeededOrBlocked": {
        "description": "Review is needed, or release has been blocked",
        "content": {
          "application/json":{
            "schema":{
              "type": "object",
              "properties": {
                "status": { "type": "string"}
              },
              "example": {
                "error": "Review needed"
              }
            }
          }
        }
      },
      "MultipleReviewError": {
        "description": "Only one review per task is allowed",
        "content": {
          "application/json":{
            "schema":{
              "type": "object",
              "properties": {
                "error": { "type": "string"}
              },
              "example": {
                "error": "Task has been already reviewed"
              }
            }
          }
        }
      },
      "NotFound": {
        "description": "Not found."
      },
      "InvalidBody": {
        "description": "The request body cannot be parsed or is missing required values."
      },
      "Unauthenticated": {
        "description": "Unauthenticated."
      },
      "Unauthorized": {
        "description": "Unauthorized."
      },
      "InternalError": {
        "description": "An internal error occurred."
      },
...
      "Conflict": {
        "description": "Conflict. Resource already exists",
        "content": {
          "application/json":{
            "schema":{
              "type": "object",
              "properties": {
                "error": { "type": "string"}
              },
              "example": {
                "error": "Pod is already running"
              }
            }
          }
        }
      },
...
    },
    "schemas":{
...
      "TaskPostBody": {
        "type": "object",
        "properties": {
          "name": { "type": "string", "maxLength": 256},
          "description": { "type": "string", "maxLength": 4096},
          "executors": {
            "type": "object",
            "properties": {
              "image":{
                "type": "string", "maxLength": 256
              },
              "command":{
                "type": "array",
                "items":{
                  "type": "string"
                }
              },
              "env":{
                "type": "object",
                "example": {
                  "LOGS": "debug",
                  "PARAMETER1": 2
                }
              }
            },
            "required": ["image"]
          },
          "tags": {
            "type": "object",
            "description": "Either dataset_id or dataset_name has to be set",
            "properties": {
              "dataset_id":{ "type": "string" },
              "dataset_name":{ "type": "string" }
            },
            "example": {
              "dataset_id": 1,
              "custom_tag": "value"
            },
            "required": ["dataset_id"]
          },
          "outputs": {
            "type": "object",
            "description": "An object where the key is a human readable name and the value is the path to which a volume will be mounted on at run time to fetch the results. The key value, in the example folder1, will be used as folder name in the results archive as well.",
            "example": {
              "folder1": "/mnt/data/results",
              "folder2": "/home/john/analysis/csvs"
            }
          },
          "db_query": {
            "type": "object",
            "properties": {
              "query": {
                "type": "string",
                "description": "The intended query to run",
                "example": "SELECT AVERAGE(pressure) FROM patients_blood;"
              },
              "dialect": {
                "type": "string",
                "description": "The type of query engine is written for, just so it can be accurately converted to the hosts' local DB",
                "enum": [
                  "postgres",
                  "mssql",
                  "oracle",
                  "mysql",
                  "mariadb"
                ]
              }
            }
          },
          "resources": {
            "type": "object",
            "properties": {
              "cpu_cores": { "type": "integer" },
              "preemptible": { "type": "boolean" },
              "ram_gb": { "type": "integer" },
              "disk_gb": { "type": "integer" },
              "zones": { "type": "string" }
            }
          },
          "volumes": {
            "type": "array",
            "items": { "type": "string" }
          }
        },
        "required": ["executors", "name", "tags"]
      },
      "TaskById": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string",
            "example": "1"
          },
          "created_at": {
            "type": "string",
            "example": "Mon, 19 Feb 2024 10:41:54 GMT"
          },
          "description": {
            "type": "string",
            "example": "Custom task description"
          },
          "docker_image": {
            "type": "string",
            "example": "registry/image:1.2"
          },
          "dataset_id": {
            "type": "integer",
            "example": 1
          },
          "requested_by": {
            "type": "string",
            "example": "aabf657d-e912-43ab-8df7-924731b0f74f"
          },
          "status": {
            "type": "object",
            "properties": {
              "running": {"type": "object"},
              "terminated": {"type": "object"},
              "waiting": {"type": "object"}
            }
          },
          "title": {
            "type": "string",
            "example": "Average age analysis"
          },
          "updated_at": {
            "type": "string",
            "example": "Mon, 19 Feb 2024 11:41:54 GMT"
          }
        }
      },
...
    }
  }
}